Fixed bug where pasting a game didn't update the whosturn text on the bottom bar
